+++ src/main/java/atdd/path/application/dto/Station/CreateStationRequestView.java
@@ -1,4 +1,4 @@
-package atdd.path.application.dto;
+package atdd.path.application.dto.Station;
 
 import atdd.path.domain.Station;
