+++ src/main/java/atdd/path/dao/UserDao.java
@@ -1,6 +1,6 @@
 package atdd.path.dao;
 
-import atdd.path.application.dto.FindByEmailResponseView;
+import atdd.path.application.dto.User.FindByEmailResponseView;
 import atdd.path.application.exception.NoDataException;
 import atdd.path.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
