+++ src/test/java/atdd/path/web/LineHttpTest.java
@@ -1,6 +1,6 @@
 package atdd.path.web;
 
-import atdd.path.application.dto.LineResponseView;
+import atdd.path.application.dto.Line.LineResponseView;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
 import org.springframework.test.web.reactive.server.WebTestClient;
